UI changes for fx.ChooseAFamily.
Now all the families in the game are returned in alphabetical order, and the pop-up contains a search box that filters a  selectable list of families.

It is NOT cancellable by default, because there are no cards that have the "Choose a race" action cancellable, i.e. it is mandatory to choose a race in the effect.